refactor(net): split lib.rs to satisfy the file-size gate
Extract the NaN-box value-conversion helpers (string_from_header_i64,
jsvalue_to_socket_bytes, is_nanboxed_pointer, unbox_pointer, the
get_object_{string,number,bool}_field readers, and build_error_object)
plus their perry-runtime extern declarations into a new sibling module
jsvalue.rs, re-exported at the crate root so every crate::<fn> call site
in lib.rs and the tls/classes/ip/lifecycle/option_setters siblings is
unchanged. Pure code movement, no logic change. Brings lib.rs from 2022
to 1824 lines, under the 2000-line gate.
